Move measurement state out of TimelineFocusManager into a per-TraceView controller.
Measurement state was living on the TimelineFocusManager singleton, which would
have leaked state between multiple loaded traces. Extract it into a new
MeasurementController class owned by TraceView and threaded through TimelineView
and FlameTrackItem. The drag-target slot also moves off a function-local static
and onto TimelineView as a typed enum.
