+++ CMakeLists.txt
@@ -76,6 +76,7 @@ set(VIEW_FILES
     src/view/src/rocprofvis_hotkey_manager.cpp 
     src/view/src/rocprofvis_stickynote.cpp
     src/view/src/rocprofvis_click_manager.cpp
+    src/view/src/rocprofvis_measurement_controller.cpp
     src/view/src/rocprofvis_analysis_view.cpp 
     src/view/src/rocprofvis_trace_view.cpp 
     src/view/src/rocprofvis_line_track_item.cpp
