+++ src/view/src/rocprofvis_click_manager.h
@@ -2,9 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #pragma once
-#include <cstdint>
 #include <map>
-#include <string>
 #include <vector>
 namespace RocProfVis
 {
@@ -18,32 +16,6 @@ enum class Layer
     kInteractiveLayer,
     kScrubberLayer,
     kCount
-};
-
-enum class MeasurementState
-{
-    kInactive,
-    kWaitingForFirst,
-    kWaitingForSecond,
-    kComplete
-};
-
-enum class MeasureEdge
-{
-    kStart,
-    kEnd
-};
-
-struct MeasurementPoint
-{
-    double      timestamp = 0.0;
-    double      duration  = 0.0;
-    uint64_t    track_id  = 0;
-    uint32_t    level     = 0;
-    std::string name;
-    uint64_t    event_uuid = 0;
-    bool        valid      = false;
-    bool        freehand   = false;
 };
 
 class TimelineFocusManager
@@ -60,43 +32,15 @@ public:
     Layer GetRightClickLayer() const;
     void  ClearRightClickLayer();
 
-    // Measurement mode
-    void             EnterMeasurementMode();
-    void             ExitMeasurementMode();
-    bool             IsMeasurementMode() const;
-    MeasurementState GetMeasurementState() const;
-
-    void SetMeasurementPoint(double timestamp, double duration,
-                             uint64_t track_id, uint32_t level,
-                             const std::string& name, uint64_t event_uuid);
-    void SetFreehandMeasurementPoint(double timestamp);
-    void ClearMeasurement();
-
-    const MeasurementPoint& GetPoint(int index) const;
-    MeasureEdge             GetEdge(int index) const;
-    void                    SetEdge(int index, MeasureEdge edge);
-    double                  GetEffectiveTimestamp(int index) const;
-
-    void   SetFreehandMode(bool enabled);
-    bool   IsFreehandMode() const;
-    void   SetFreehandOffset(int index, double offset_ns);
-    double GetFreehandOffset(int index) const;
-
 private:
     TimelineFocusManager();
-    ~TimelineFocusManager()                              = default;
+    ~TimelineFocusManager()                                      = default;
     TimelineFocusManager(const TimelineFocusManager&)            = delete;
     TimelineFocusManager& operator=(const TimelineFocusManager&) = delete;
 
     Layer                 m_layer_focused;
     std::map<Layer, bool> m_all_layers_focused;
     Layer                 m_right_click_layer;
-
-    MeasurementState m_measurement_state;
-    MeasurementPoint m_points[2];
-    MeasureEdge      m_edges[2];
-    bool             m_freehand_mode;
-    double           m_freehand_offsets[2];
 };
 
 }  // namespace View
